# apps/shared/fastapi_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "coding-agent backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks/{task_id}/turns": {
      "get": {
        "summary": "List Turns",
        "description": "列出某任务下的全部 turn（支撑多轮历史展示）。\n\n参数:\n    task_id: 来自路由的任务标识。\n    turn_service: 通过依赖注入的轮次 service。\n\n返回:\n    该任务下按创建时间升序的 ``TurnResponse`` 列表。\n\n异常:\n    HTTPException: 当任务不存在（级联 KeyError）时抛出。\n\n副作用:\n    无。",
        "operationId": "list_turns_tasks__task_id__turns_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurnResponse"
                  },
                  "title": "Response List Turns Tasks  Task Id  Turns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Turn",
        "description": "为已有任务追加一个 pending 轮次。\n\n该端点只负责把用户本轮输入持久化为一个新的、处于 ``pending`` 状态的\n``Turn`` 记录，并立即返回其 ``turn_id``；**不会**触发 Agent 运行。\n真正的执行由客户端随后对 ``GET /turns/{turn_id}/stream`` 的订阅触发，\n这样客户端可以先拿到 ``turn_id`` 再建立 SSE 连接，避免竞态。\n\n参数:\n    task_id: 来自路由的任务标识。\n    payload: 包含本轮用户输入文本的请求体（``input_text``）。\n    turn_service: 通过依赖注入的轮次 service。\n\n返回:\n    创建后的 ``TurnResponse``，包含 ``turn_id``、所属 ``task_id``、状态与\n    创建时间等字段，供客户端建立后续 SSE 流使用。\n\n异常:\n    HTTPException: 当任务不存在（404）、模型不支持视觉（422）或输入非法（400）时抛出。\n\n副作用:\n    在存储中创建 turn 记录，但不启动运行、不产生运行时事件。",
        "operationId": "create_turn_tasks__task_id__turns_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/turns/{turn_id}/stream": {
      "get": {
        "summary": "Stream Turn",
        "description": "通过 SSE 流式返回轮次的运行时事件。\n\n仅对 ``pending`` 状态的轮次启动 Agent 运行并实时推送事件。非 pending\n轮次由调用方 409 守卫拒绝，历史回看请走 ``GET /tasks/{task_id}/turns``。\n本端点只负责「执行」，不承担回放或断线重连（断开即本轮结束，由运行时\n``run_turn`` 的断开兜底标 failed）。本次运行使用的 agent 在轮次创建时即已绑定\n（``turn.agent_id``），\n运行时按「turn 绑定 > task 默认」解析，无需本端点再传参。\n\n事件以标准 SSE 帧推送，每帧格式为::\n\n    event: <event_type>\n    data: <json 序列化后的 RuntimeEvent>\n\n其中 ``event_type`` 对应 ``RuntimeEvent`` 的种类（如 ``token``、\n``tool_call``、``step``、``status`` 等），``data`` 为该事件的 JSON 负载。\n相邻事件以空行 ``\\n\\n`` 分隔。\n\n参数:\n    turn_id: 来自路由的轮次标识。\n    runtime: 通过依赖注入的运行时（仅用于执行 pending 轮次）。\n    turn_service: 通过依赖注入的轮次 service（用于取轮次记录与状态守卫）。\n    stream_service: 通过依赖注入的轮次事件流编排 service（订阅/认领/驱动/发布/兜底）。\n\n返回:\n    发送 ``text/event-stream`` 的 StreamingResponse，连接保持打开直到\n    轮次运行结束或客户端断开（断开即本轮结束，由运行时 ``run_turn`` 的断开兜底标 failed）。\n\n异常:\n    HTTPException: 当轮次不存在（404）或轮次非 pending（409，历史请走\n        ``GET /tasks/{task_id}/turns``）时抛出。\n\n副作用:\n    仅对 pending turn 调用 ``runtime.run_turn`` 启动运行；事件产出由\n    ``_sse_frames`` 转换为 SSE 帧。不回放历史事件。",
        "operationId": "stream_turn_turns__turn_id__stream_get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/turns/{turn_id}/events/stream": {
      "get": {
        "summary": "Subscribe Turn Events",
        "description": "订阅已运行委派子轮次的实时事件。\n\n参数:\n    turn_id: 来自路由的子轮次标识。\n    turn_service: 用于读取轮次并校验其执行状态的领域 service。\n    stream_service: 通过依赖注入的轮次事件流编排 service（纯订阅转发）。\n\n返回:\n    ``text/event-stream`` 响应，仅转发此后实时发布到该 turn 的事件。\n\n异常:\n    HTTPException: turn 不存在时为 404；turn 已终态时为 409。\n\n副作用:\n    请求消费期间注册事件订阅；不认领 producer、不调用运行时，也不在断开时更新 turn 状态。",
        "operationId": "subscribe_turn_events_turns__turn_id__events_stream_get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/turns/{turn_id}/cancel": {
      "post": {
        "summary": "Cancel Turn",
        "description": "取消指定轮次并中止其运行。\n\n置 turn 为 ``cancelled``，模型节点在下一轮循环检查到取消状态后停止派发工具，从而中止运行。\n\n参数:\n    turn_id: 来自路由的轮次标识。\n    runtime: 通过依赖注入的运行时（负责取消与状态推进）。\n    turn_service: 通过依赖注入的轮次 service（用于校验 turn 存在）。\n\n返回:\n    取消后的 ``TurnResponse``。\n\n异常:\n    HTTPException: 当轮次不存在（404）时抛出。\n\n副作用:\n    置 turn 为 cancelled 并 emit 取消事件。",
        "operationId": "cancel_turn_turns__turn_id__cancel_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentRef": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "file",
              "directory",
              "url"
            ],
            "title": "Kind"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ref"
        ],
        "title": "AttachmentRef"
      },
...
      "CreateTurnRequest": {
        "properties": {
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AttachmentRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          }
        },
        "type": "object",
        "required": [
          "input_text"
        ],
        "title": "CreateTurnRequest",
        "description": "校验轮次创建请求体。\n\n参数:\n    input_text: 非空的本轮用户输入。\n    provider_id: 可选，模型归属的厂商标识（指向 ``providers.id``）；None 表示未选厂商。\n    model_name: 可选，本 turn 关联的模型名；None 表示未选模型。\n    attachments: 可选，本 turn 携带的附件引用列表（结构化，含类型与引用）。\n        仅做结构合法性校验；workspace 边界、视觉能力等业务规则在 service 层判定。\n    reasoning_effort: 可选，本 turn 思考努力等级；None 表示 max。该值经\n        ``turn_service.create_turn`` 透传落库到 ``turns.reasoning_effort``。\n\n返回:\n    Pydantic 请求模型。\n\n异常:\n    ValueError: 当输入为空白或附件列表非法时抛出。\n\n副作用:\n    无。本层只做结构校验，不访问文件系统、不做业务规则判定。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TurnResponse": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "end_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Reason"
          },
          "response_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "turn_id",
          "task_id",
          "input_text",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TurnResponse",
        "description": "校验并序列化轮次状态响应（与 ``TurnRecord`` 字段对齐）。\n\n参数:\n    turn_id: 轮次标识。\n    task_id: 所属任务标识。\n    input_text: 本轮用户输入。\n    status: 轮次状态。\n    end_reason: 终态原因，可能为 None。\n    response_text: Agent 回复文本，可能为 None。\n    agent_id: 驱动该轮次的 agent 标识，可能为 None。\n    model_name: 该轮实际所用模型名（解析后落库，D1）；历史 turn 或运行期尚未\n        回写的轮次为 None（不表示 Auto，仅表示未落库）。\n    created_at: 创建时间文本。\n    updated_at: 更新时间文本。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}